build: handle missing error

+++ scripts/copy-target.mjs
@@ -1,6 +1,7 @@
-import fs from 'fs';
+import fs, { existsSync } from 'fs';
 import path from 'path';
 import { fileURLToPath, URL } from 'url';
+import { exit } from 'process';
 
 import { getModuleName } from './binaries.mjs';
 
@@ -14,5 +15,10 @@ if (!fs.existsSync(lib)) {
 const source = path.join(__dirname, '..', 'build', 'Release', 'sentry_cpu_profiler.node');
 const target = path.join(__dirname, '..', 'lib', getModuleName());
 
-console.log('Renaming', source, 'to', target);
-fs.renameSync(source, target);
+if (!existsSync(source)) {
+  console.log('Source file does not exist:', source);
+  exit(1);
+} else {
+  console.log('Renaming', source, 'to', target);
+  fs.renameSync(source, target);
+}
